saas: remove notes, devcontainer
