--- starters/saas/.devcontainer/.env
@@ -1,6 +1,0 @@
-POSTGRES_DB=loco_app
-POSTGRES_USER=loco
-POSTGRES_PASSWORD=loco
-DATABASE_URL=postgres://loco:loco@db:5432/loco_app
-REDIS_URL=redis://redis:6379
-MAILER_HOST=mailer
