--- starters/saas/.devcontainer/Dockerfile
@@ -1,8 +1,0 @@
-FROM mcr.microsoft.com/devcontainers/rust:1
-
-RUN apt-get update && export DEBIAN_FRONTEND=noninteractive \
-     && apt-get -y install --no-install-recommends postgresql-client \
-     && cargo install sea-orm-cli cargo-insta \
-     && chown -R vscode /usr/local/cargo
-
-COPY .env /.env
