--- starters/saas/.devcontainer/devcontainer.json
@@ -1,9 +1,0 @@
-{
-    "name": "Loco",
-    "dockerComposeFile": "compose.yaml",
-    "service": "app",
-    "workspaceFolder": "/workspaces/${localWorkspaceFolderBasename}",
-    "forwardPorts": [
-        5150
-    ]
-}
