+++ starters/saas/migration/src/lib.rs
@@ -3,16 +3,12 @@
 pub use sea_orm_migration::prelude::*;
 
 mod m20220101_000001_users;
-mod m20231103_114510_notes;
 
 pub struct Migrator;
 
 #[async_trait::async_trait]
 impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
-        vec![
-            Box::new(m20220101_000001_users::Migration),
-            Box::new(m20231103_114510_notes::Migration),
-        ]
+        vec![Box::new(m20220101_000001_users::Migration)]
     }
 }
