--- starters/saas/migration/src/m20231103_114510_notes.rs
@@ -1,34 +1,0 @@
-use loco_rs::schema::table_auto_tz;
-use sea_orm_migration::{prelude::*, schema::*};
-
-#[derive(DeriveMigrationName)]
-pub struct Migration;
-
-#[async_trait::async_trait]
-impl MigrationTrait for Migration {
-    async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .create_table(
-                table_auto_tz(Notes::Table)
-                    .col(pk_auto(Notes::Id))
-                    .col(string_null(Notes::Title))
-                    .col(string_null(Notes::Content))
-                    .to_owned(),
-            )
-            .await
-    }
-
-    async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .drop_table(Table::drop().table(Notes::Table).to_owned())
-            .await
-    }
-}
-
-#[derive(DeriveIden)]
-enum Notes {
-    Table,
-    Id,
-    Title,
-    Content,
-}
