+++ starters/saas/src/controllers/mod.rs
@@ -1,3 +1,1 @@
 pub mod auth;
-pub mod notes;
-pub mod user;
