+++ starters/saas/src/models/_entities/mod.rs
@@ -2,5 +2,4 @@
 
 pub mod prelude;
 
-pub mod notes;
 pub mod users;
