+++ starters/saas/src/models/_entities/prelude.rs
@@ -1,4 +1,3 @@
 //! `SeaORM` Entity, @generated by sea-orm-codegen 1.0.0
 
-pub use super::notes::Entity as Notes;
 pub use super::users::Entity as Users;
