+++ starters/saas/src/models/mod.rs
@@ -1,3 +1,2 @@
 pub mod _entities;
-pub mod notes;
 pub mod users;
