--- starters/saas/src/models/notes.rs
@@ -1,7 +1,0 @@
-use sea_orm::entity::prelude::*;
-
-use super::_entities::notes::ActiveModel;
-
-impl ActiveModelBehavior for ActiveModel {
-    // extend activemodel below (keep comment for generators)
-}
