+++ starters/saas/src/views/mod.rs
@@ -1,2 +1,1 @@
 pub mod auth;
-pub mod user;
