--- starters/saas/src/views/user.rs
@@ -1,21 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::models::_entities::users;
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct CurrentResponse {
-    pub pid: String,
-    pub name: String,
-    pub email: String,
-}
-
-impl CurrentResponse {
-    #[must_use]
-    pub fn new(user: &users::Model) -> Self {
-        Self {
-            pid: user.pid.to_string(),
-            name: user.name.clone(),
-            email: user.email.clone(),
-        }
-    }
-}
