+++ starters/saas/tests/requests/mod.rs
@@ -1,4 +1,2 @@
 mod auth;
-mod notes;
 mod prepare_data;
-mod user;
